tidy: move ColumnKind to its own file
timeline file is starting to get messy
