+++ src/app.rs
@@ -3,6 +3,7 @@ use crate::actionbar::BarResult;
 use crate::app_creation::setup_cc;
 use crate::app_style::user_requested_visuals_change;
 use crate::args::Args;
+use crate::column::ColumnKind;
 use crate::draft::Drafts;
 use crate::frame_history::FrameHistory;
 use crate::imgcache::ImageCache;
@@ -12,11 +13,11 @@ use crate::notecache::{CachedNote, NoteCache};
 use crate::relay_pool_manager::RelayPoolManager;
 use crate::route::Route;
 use crate::thread::{DecrementResult, Threads};
-use crate::timeline::{ColumnKind, Timeline, TimelineSource, ViewFilter};
+use crate::timeline::{Timeline, TimelineSource, ViewFilter};
 use crate::ui::note::PostAction;
 use crate::ui::{self, AccountSelectionWidget, DesktopGlobalPopup};
 use crate::ui::{DesktopSidePanel, RelayView, View};
-use crate::{Result, filter};
+use crate::{filter, Result};
 use egui_nav::{Nav, NavAction};
 use enostr::{ClientMessage, RelayEvent, RelayMessage, RelayPool};
 use std::cell::RefCell;
